Moved config data sources package. Set timeout for firebase RC
